Avoid wasting memory for unused print layout positions, by instantiating
2007-10-18  Murray Cumming  <murrayc@murrayc.com>

* glom/libglom/data_structure/layout/layoutitem.cc:
Avoid wasting memory for unused print layout positions, by 
instantiating an inner class on demand.
* glom/mode_design/print_layouts/canvas_print_layout.cc:
* glom/mode_design/print_layouts/canvas_print_layout.h: 
Add edit and formatting context menu items.
Inherit from Base_DB so it can access the document.
Add get/set_page_setup().
* glom/mode_design/print_layouts/window_print_layout_edit.cc:
* glom/mode_design/print_layouts/window_print_layout_edit.h:
Add the canvas as a view so it can access the document.
Add a page setup menu to modify the page setup of the canavs.
* glom/utility_widgets/canvas/canvas_group_resizable.cc:
* glom/utility_widgets/canvas/canvas_group_resizable.h:
Do not sigc::bind the RefPtr manipulator items in the signal handler.
This seems to cause some strange memory problems. I already checked 
that this is not caused by the multiple virtual inheritance. This 
was unpleasant.

svn path=/branches/glom-precise-printing/; revision=1225