Added get_print_layout_text_size(), though we should have this for
2007-10-23  Murray Cumming  <murrayc@murrayc.com>

* glom/libglom/data_structure/layout/layoutitem.cc:
* glom/libglom/data_structure/layout/layoutitem.h: 
Added get_print_layout_text_size(), though we should have this for 
on-screen text too.
* glom/utility_widgets/canvas/canvas_text_movable.cc:
* glom/utility_widgets/canvas/canvas_text_movable.h: Added set_text() 
and set_text_size(), so we can use pango markup.
* glom/mode_design/print_layouts/canvas_layout_item.cc: set_layout_item(): 
Set the text size from the layout item.
* glom/mode_design/print_layouts/canvas_print_layout.cc:
* glom/mode_design/print_layouts/canvas_print_layout.h: Constructor: 
Set the units to mm, and set the page setup to a default.
set_page_setup(): Change the canvas bounds to match the page setup. 
* glom/mode_design/print_layouts/window_print_layout_edit.cc:
* glom/mode_design/print_layouts/window_print_layout_edit.h:
init_menu(): Added zoom items to the view menu, and start with 50%.

svn path=/trunk/; revision=1234