get_table_names_from_database(): Comment out the checks for internal
2009-02-06  Murray Cumming  <murrayc@murrayc.com>

* glom/base_db.cc: get_table_names_from_database(): Comment out the 
checks for internal tables, because that is now fixed in libgda.
* glom/mode_data/box_data.[h|cc]:
* glom/base_db_table_data.[h|cc]: Move refresh_related_fields() 
(which is virtual) and get_related_fields() down the hierarchy so that 
the override is also used by DbAddDel, so related fields are updated 
in the view on list views and related records.
This really fixes bug #569722 (Johannes Schmid)

svn path=/trunk/; revision=1881