Refactoring Dialog_Import_CSV: created testcase for the CsvParser's signal emission
* Makefile_tests.am: all tests now run within valgrind by default since mem
leaks in testcases are potential mem leaks in the application. Parameters might
need tweaking.

* glom/import_csv.cc (CsvParser::on_idle_parse): Fixed a missing negation in a
conditional, which was wrongfully leading to signal emissions. Found by the
import signals testcase.

* test/import/test_parsing.cc: Fixed a null byte issue when setting the raw
contents of the parser.

* test/import/test_signals.cc: This testcase checks the CsvParser's signal
emission on varying input. The basic idea was to have on testcase for each
code path that leads to a signal emission in glom/import_csv.cc, even if that
leads to overlapping code coverage tests with the parser testcase.