Maemo: List: Click to view details.
* glom/utility_widgets/db_adddel/db_adddel.[h|cc]: Connect to
TouchSelector::signal_changed to respond to clicks on the list. Note that
a drag to pan (scroll) does not cause that signal to be emitted.
* glom/utility_widgets/db_adddel/db_adddel_withbuttons.[h|cc]: On Mameo,
remove the Delete and Edit buttons because deleting can be done via the
details window (viewable on Maemo). The Add button is still there and is ugly.